story-13 refactor class name
