+++ Silkrode/app/src/main/java/personal/ivan/silkrode/navigation/podcast/di/PodcastScopeDi.kt
@@ -13,7 +13,7 @@ import personal.ivan.silkrode.db.PodcastDao
 import personal.ivan.silkrode.di.ViewModelKey
 import personal.ivan.silkrode.navigation.podcast.view.PodcastActivity
 import personal.ivan.silkrode.navigation.podcast.view.fragment.PlayFragment
-import personal.ivan.silkrode.navigation.podcast.view.fragment.collection_list.CollectionFragment
+import personal.ivan.silkrode.navigation.podcast.view.fragment.collection.CollectionFragment
 import personal.ivan.silkrode.navigation.podcast.view.fragment.pod_cast_list.PodcastListFragment
 import personal.ivan.silkrode.navigation.podcast.viewmodel.PodcastViewModel
 import personal.ivan.silkrode.util.DateFormatUtil
